@Bilge 2025-02-14
`xdebug.start_with_request=yes` is an anti-pattern in any case. Developers need to learn how to use triggers.