@carlos-granados 2025-02-15
@Bilge @nickchomey this will benefit you as well if you use `xdebug.start_with_request=trigger`. Currently Xdebug will slow down your code even if you use `trigger` and don't trigger the debugger