@marcosgrimm 2025-02-18
@Bilge > `xdebug.start_with_request=yes` is an anti-pattern in any case. Developers need to learn how to use triggers.

There would be a need to trigger for each different type of request, no?  From cli, from postman, from web... each one should have a way to trigger it. Is that what you are saying? :) 

For me, I always use `xdebug.start_with_request=yes` in my config, but it would be nice to make it only work with the triggers...