@derickr 2025-02-18
This seems to be a massive change, which hasn't been discussed up-front. Please reach out first through the mailing list, as per the [Contributing Guidelines](https://github.com/xdebug/xdebug/blob/master/CONTRIBUTING.rst). It also seems to be changing behaviour, and mixes a few things together, which really ought to be different pull requests.

I also don't appreciate sponsor links in Pull Requests, so I will remove that from the description.

But as I said, reach out first.