@carlos-granados 2025-02-18
@derickr no problem about the sponsor link, did not know you did not like that. No problem splitting this as needed. You said to reach out through the mailing list. Which mailing list? Can you please be a little more specific? Thanks!