@carlos-granados 2025-03-01
While testing this I realised that in some cases this would not work correctly as it was enabling/disabling the statement handler for all requests. I updated the code with a new implementation that does not fully disable this handler and instead adds a check to see it the debugger is active. The new code is a little less performant, as expected, but still provides a really good improvement. These are the new timing values:

| **Operation**            | **Without Xdebug (s)** | **With Current Xdebug (s)** | **Slowdown (%)** | **With Optimized Xdebug (s)** | **Slowdown (%)** |
|--------------------------|----------------------|----------------------|----------------|----------------------|----------------|
| Running **RectorPHP**    | 42                   | 92                   | **+119%**       | **59**                | **+40%**       |
| Running **PHPStan**      | 51                   | 108                  | **+112%**       | **69**                | **+35%**       |