@TwistedAndy 2025-03-21
It would be great to see this PR merged, because now I have to constantly enable and disable Xdebug during development to get better performance.