@derickr 2025-04-11
Sorry it took some time, but there is a lot of it in here. Large PRs that do a
multitude of things are hard to review.

From what I understand, the main point is that this only improves performance
when the debugger isn't connected, but activated in its mode? Wouldn't it be
easier to not turn it on (in modes) to begin with?

I'm sure a few of the other changes are still useful, but there are too many
things happening in one PR to reasonable think about to see the effective
changes.

Mostly, because it tries to do too many things at once.

Architecture issues:

- 'base' and 'lib' should not access XG_DBG.
- not a fan of having flags named negatively (disabled, false, etc)
- boolean flags (to `xdebug_add_stack_frame`) should be `bool`, not `int`.
- the term `debugger_disabled` is confusing. The user did not disable the
  debugger. Instead, some heuristics decided not to turn it on for this
  request.

Specific code issues:

- `xdebug_add_stack_frame` was modified with an additional i(`int`) argument.
  Instead of that, it should instead always accept for which
  `zend_execute_data` to add a stack frame for, without switching between
  `EG(current_execute_data)` and `zdata` — this can just be a first PR with
  just this change.

- Rebuilding the stack frames recursively might not work if fibers are
  involved. I see no tests that try to validate this.

- `xdebug_enable_debugger_if_disabled` / `xdebug_disable_debugger_if_enabled`:
  should be a boolean flag, but it also makes no sense to check for the
  current value first, as all you do is just setting it to the other. That can
  be an unconditional set. Checking with a condition first doesn't make that
  process go faster.

- In `src/base/base.h`, I don't understand the inclusion of
  `zend_extensions.h`.

- In `src/base/ctrl_socket.c` you add a call to
  `xdebug_enable_debugger_and_rebuild_stack_if_disabled();` — wouldn't it make
  more sense to do this (for all cases) when Xdebug reads the
  `do_connect_to_client` flag and acts on it?

- In `src/debugger/com.c` you check for `XDEBUG_MODE_IS_ONLY_DEBUG` - that
  seems to be superfluous.

- In `xdebug_opcode_multi_handler` — overridden opcode handles are also
  useful for other modes, not just for the step debugger. I worry that this
  makes problems. In any case, that change could be in it's own PR.

- Multi-overridden opcodes are only done when debugger is disabled, this
  is not likely correct, as these are also used in other cases. And also,
  does it really make a big difference?

- Changes how non-connectivity gets reported in `xdebug_init_debugger`, from
  `XLOG_ERR` to `XLOG_WARN` — that doesn't need to be in the
  same PR, and might not be wanted.

- In tests, such as in `tests/debugger/bug01656.phpt`, the change should be a
  utility function, as this is now used loads of times.

- In `xdebug.c` you only call `xdebug_library_minit` conditionally. I've tried
  hard in the past to make this code not spaghetti, and this makes it
  spaghetti again. I'm sure there is a neater way of doing this.

- in `xdebug_statement_call`, the IS_OFF should be checked first. But
  statement calls are also used with code coverage, which means that if
  the debugger was 'disabled', code coverage would not work either now, as
  the function gets short-circuited now.

- `xdebug_init_oparray`'s 'optimisation' — it should have the check
  around the `xdebug_coverage_init_oparray()` function itself, just like in
  ZEND_MODULE_POST_ZEND_DEACTIVATE_D — this should also be a PR just on its
  own.

In short, break this PR up in many little ones. One doing a simple thing, and
only one at a time. Big PRs like this get confusing quickly, and therefore
will unlikely get merged.

I'm closing this PR now, and am looking forwards to seeing PRs with
incremental improvements.