@tanerkay 2025-04-11
> From what I understand, the main point is that this only improves performance when the debugger isn't connected, but activated in its mode? Wouldn't it be easier to not turn it on (in modes) to begin with?

I agree with your assessment about this PR, and that it should be split up. But I feel this main point is an important one. As @TwistedAndy mentions, it's a pain enabling/disabling xdebug, even after creating scripts that replace lines/files in the config and do any needed restarting of services.