@TwistedAndy 2025-04-11
Unfortunately, phpdismod and ddev is not an option for a lot of cases and server setups.

Ultimately, I have ended up with writing a script which changes the Xdebug mode on the active PHP version and restarts a server. Yes, it's a workaround and has a lot of restrictions, but it's better than editing php.ini manually EVERY time I want to debug something.

Even with this workaround I find myself using **print_r()** and **dd()** just because it's faster.