@staabm 2025-04-11
I am using https://github.com/w00fz/xdebug-osx to enable xdebug on/off 

Agree its too slow to have it permanently on and toggling it with a cli script is the best I could come up with